requirements.txt
Django==4.2.5
djangorestframework==3.14.0
PyJWT==2.6.0
python-dotenv==1.1.0
pymongo==4.11.3
redis==4.5.5
celery==5.3.4
django-cors-headers==3.14.0
django-filter==23.2
django-environ==0.10.0
gunicorn==21.2.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
django-storages==1.13.2
boto3==1.28.0
Werkzeug==3.1.3
django-celery-results==2.5.1